+++ playground/index.ts
@@ -4,7 +4,7 @@ import { createApp, createRouter } from '../src'
 const app = createApp()
 const router = createRouter()
   .get('/', () => 'Hello World!')
-  .get('/hello/:name', req => `Hello ${req.params.name}!`)
+  .get('/hello/:name', event => `Hello ${event.params.name}!`)
 
 app.use(router)