+++ src/handler.ts
@@ -2,8 +2,10 @@ import { withoutTrailingSlash, withoutBase } from 'ufo'
 import type { Handler, PromisifiedHandler, Middleware, IncomingMessage, ServerResponse, LazyHandler } from './types'
 
 export const defineHandler = <T>(handler: Handler<T>) => handler
+
 /** @deprecated Use defineHandler */
 export const defineHandle = defineHandler
+
 export const defineMiddleware = (middleware: Middleware) => middleware
 
 export function promisifyHandler (handler: Handler | Middleware): PromisifiedHandler {
@@ -11,6 +13,9 @@ export function promisifyHandler (handler: Handler | Middleware): PromisifiedHan
     return callHandler(handler, req, res)
   }
 }
+
+/** @deprecated Use defineHandler */
+export const promisifyHandle = promisifyHandler
 
 export function callHandler (handler: Middleware, req: IncomingMessage, res: ServerResponse) {
   return new Promise((resolve, reject) => {
@@ -23,7 +28,7 @@ export function callHandler (handler: Middleware, req: IncomingMessage, res: Ser
   })
 }
 
-export function defineLazyHandler (handler: LazyHandler, promisify?: boolean): PromisifiedHandler {
+export function defineLazyHandler (handler: LazyHandler, promisify?: boolean): Handler {
   let _promise: Promise<Handler>
   const resolve = () => {
     if (!_promise) {
@@ -37,12 +42,10 @@ export function defineLazyHandler (handler: LazyHandler, promisify?: boolean): P
   }
 }
 
-/**
- * @deprecated Use new events API or defineLazyHandler
- */
+/** @deprecated Use defineLazyHandler */
 export const lazyHandle = defineLazyHandler
 
-export function useBase (base: string, handler: PromisifiedHandler): PromisifiedHandler {
+export function useBase (base: string, handler: Handler): Handler {
   base = withoutTrailingSlash(base)
   if (!base) { return handler }
   return function (req, res) {